pyproject.toml
[project]
name = "demo-representation"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrejev Denis",email = "andrejevdenis00@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "selene (==2.0.0rc9)",
    "allure-pytest (>=2.13.5,<3.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "appium-python-client (>=4.5.0,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "selenium (>=4.28.1,<5.0.0)",
    "allure-python-commons (>=2.13.5,<3.0.0)",
    "pycparser (>=2.22,<3.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "jsonschema (>=4.23.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = 1
log_level = "INFO"
addopts = "--clean-alluredir --alluredir=allure-results -l"
